Move responsability to find NTP servers to the Allin1 dialog
